LayoutTests/fullscreen/video-inside-flex-item-expected.txt


EVENT(webkitfullscreenchange)
TEST(video.clientWidth == document.body.clientWidth) OK
EVENT(webkitfullscreenchange)
TEST(video.clientWidth == oldVideoWidth) OK
TEST(video.clientHeight == oldVideoHeight) OK
END OF TEST
